edudap-office/requirements.txt

Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.46
python-dotenv==1.0.1
Werkzeug==3.1.5
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.3.1
psycopg2-binary==2.9.11
gunicorn==23.0.0
pandas==3.0.0
openpyxl==3.1.5
PyPDF2==3.0.1
